requirements.txt
black==24.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
greenlet==3.1.1
idna==3.10
lxml==5.3.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.38
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
